move MergeResourcesInto and StripProceduralFilters to AdBlockService
