+++ components/brave_shields/core/browser/ad_block_service_helper.h
@@ -6,24 +6,13 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_SHIELDS_CORE_BROWSER_AD_BLOCK_SERVICE_HELPER_H_
 #define BRAVE_COMPONENTS_BRAVE_SHIELDS_CORE_BROWSER_AD_BLOCK_SERVICE_HELPER_H_
 
-#include <stdint.h>
-
 #include <optional>
 #include <string>
-#include <vector>
-
-#include "base/values.h"
 
 namespace brave_shields {
 
 void MergeCspDirectiveInto(std::optional<std::string> from,
                            std::optional<std::string>* into);
-
-void MergeResourcesInto(base::Value::Dict from,
-                        base::Value::Dict& into,
-                        bool force_hide);
-
-void StripProceduralFilters(base::Value::Dict& resources);
 
 }  // namespace brave_shields
